organização da arquitetura e design do projeto

+++ src/main/java/com/stock/flex/resource/ProductResource.java
@@ -2,6 +2,8 @@ package com.stock.flex.resource;
 
 import com.stock.flex.entity.CategoryEntity;
 import com.stock.flex.entity.ProductEntity;
+import com.stock.flex.entity.StockEntity;
+import com.stock.flex.entity.UserEntity;
 import com.stock.flex.repository.CategoryRepository;
 import com.stock.flex.repository.ProductRepository;
 import com.stock.flex.resource.request.ProductRequest;
@@ -11,9 +13,12 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -30,18 +35,30 @@ public class ProductResource {
 
     @PostMapping("/{categoryId}")
     @Transactional
+    @PreAuthorize("@productPermissionChecker.hasPermission(#categoryId, principal)")
     public ResponseEntity<?> create(
             @PathVariable UUID categoryId,
-            @RequestBody ProductRequest request
-    ) {
+            @RequestBody ProductRequest request,
+            @AuthenticationPrincipal UserEntity userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
         Optional<CategoryEntity> categoryOptional = categoryRepository.findById(categoryId);
         if (categoryOptional.isEmpty()) {
-            String errorMessage = "A categoria associada ao produto não foi encontrada.";
-            return ResponseEntity.badRequest().body(errorMessage);
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("A categoria associada ao produto não foi encontrada.");
+        }
+
+        CategoryEntity category = categoryOptional.get();
+        StockEntity stock = category.getStock();
+
+        if (!stock.getUser().getId().equals(userSpringSecurity.getId())) {
+            throw new Exception("Acesso negado! Você não tem permissão para criar um produto nesta categoria.");
         }
 
         ProductEntity newProduct = new ProductEntity(request);
-        newProduct.setCategory(categoryOptional.get());
+        newProduct.setCategory(category);
 
         ProductEntity savedProduct = repository.save(newProduct);
 
@@ -50,30 +67,93 @@ public class ProductResource {
     }
 
     @GetMapping
-    public ResponseEntity<List<ProductResponse>> get() {
-        var product = repository.findAll().stream().map(ProductResponse::new).toList();
-        return ResponseEntity.ok(product);
+    public ResponseEntity<List<ProductResponse>> get(@AuthenticationPrincipal UserEntity userSpringSecurity) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
+        List<ProductEntity> products = repository.findByCategoryStockUser(userSpringSecurity);
+
+        List<ProductResponse> productResponses = products.stream().map(ProductResponse::new).toList();
+
+        return ResponseEntity.ok(productResponses);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity getById(@PathVariable UUID id) {
-        var product = repository.getById(id);
+    public ResponseEntity getById(
+            @PathVariable UUID id,
+            @AuthenticationPrincipal UserEntity userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
+        Optional<ProductEntity> productOptional = repository.findById(id);
+        if (productOptional.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Produto não encontrado.");
+        }
+
+        ProductEntity product = productOptional.get();
+
+        if (!product.getCategory().getStock().getUser().equals(userSpringSecurity)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Acesso negado! Você não tem permissão para acessar este produto.");
+        }
+
         return ResponseEntity.ok(new ProductResponse(product));
     }
 
     @PutMapping("/{id}")
     @Transactional
-    public ResponseEntity<ProductRequest> update(@PathVariable UUID id, @RequestBody ProductRequest request) {
-        var product = repository.getReferenceById(id);
-        product.updateInfo(request);
-        return ResponseEntity.ok(new ProductRequest(product));
-    }
+    public ResponseEntity<?> update(
+            @PathVariable UUID id,
+            @RequestBody ProductRequest request,
+            @AuthenticationPrincipal UserEntity userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
 
+        Optional<ProductEntity> productOptional = repository.findById(id);
+        if (productOptional.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Produto não encontrado.");
+        }
+
+        ProductEntity product = productOptional.get();
+
+        if (!product.getCategory().getStock().getUser().equals(userSpringSecurity)) {
+            throw new Exception("Acesso negado! Você não tem permissão para atualizar este produto.");
+        }
+
+        product.updateInfo(request);
+
+        ProductEntity updatedProduct = repository.save(product);
+
+        ProductResponse response = new ProductResponse(updatedProduct);
+        return ResponseEntity.ok(response);
+    }
 
     @DeleteMapping("/{id}")
     @Transactional
-    public ResponseEntity delete(@PathVariable UUID id) {
+    public ResponseEntity delete(
+            @PathVariable UUID id,
+            @AuthenticationPrincipal UserEntity userSpringSecurity
+    ) throws Exception {
+        if (Objects.isNull(userSpringSecurity)) {
+            throw new Exception("Acesso negado! O usuário não está autenticado.");
+        }
+
+        Optional<ProductEntity> productOptional = repository.findById(id);
+        if (productOptional.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Produto não encontrado.");
+        }
+
+        ProductEntity product = productOptional.get();
+
+        if (!product.getCategory().getStock().getUser().equals(userSpringSecurity)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Acesso negado! Você não tem permissão para excluir este produto.");
+        }
+
         repository.deleteById(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 }
